scripts/etl/requirements.txt
supabase>=2.0.0
shapely>=2.0.0
python-dotenv>=1.0.0
